+++ context/app/static/js/components/Detail/GlobusLink/index.js
@@ -1,0 +1,3 @@
+import GlobusLink from './GlobusLink';
+
+export default GlobusLink;
